create fieldsetcontainer component
